src/requirements.txt
fastapi~=0.135.1 # фреймворк
pydantic~=2.12.5 # валидатор значений
uvicorn~=0.41.0 # мини сервер

sqlalchemy~=2.0.48 #
psycopg2~=2.9.11 #

python-dotenv~=1.2.2 # файл с ключами, глоб. переменными